use unsigned long as function id

+++ emptypushpop/emptypushpop.c
@@ -1,4 +1,4 @@
-void _instroPushIdentifier(unsigned long long funcIdentifier);
+void _instroPushIdentifier(unsigned long funcIdentifier);
 void _instroPopIdentifier();
 
 void __cyg_profile_func_enter(void *func_ptr, void *call_site);
@@ -7,7 +7,7 @@ void __cyg_profile_func_exit(void *func_ptr, void *call_site);
 void printBenchmarkResult() {
 }
 
-void _instroPushIdentifier(unsigned long long funcIdentifier) {
+void _instroPushIdentifier(unsigned long funcIdentifier) {
 	// Nothings happens here.
 	return;
 }
